AI_server/python_version_independent.requirements.txt
django==4.2
djangorestframework
boto3
pymongo
python-dotenv
ipython
requests-toolbelt
openai
google-cloud-texttospeech
google_api_python_client
google_auth_oauthlib
pydub
pyrubberband
ffprobe-python
azure-cognitiveservices-speech
langcodes
language_data
numpy
deepl
regex
soundfile>=0.12.1
aiohttp